requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.5; python_version >= '3.6'
aniso8601==9.0.1; python_version >= '3.5'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
flask-cors==3.0.10
flask-migrate==3.1.0
flask-restx==0.5.1
flask-sqlalchemy==2.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask==2.0.2
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
jsonschema==4.2.1; python_version >= '3.7'
mako==1.1.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
psycopg2==2.9.2
pycparser==2.21
pyjwt==2.3.0
pyrsistent==0.18.0; python_version >= '3.6'
pytz==2021.3
setuptools==59.5.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.27; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.0.2; python_version >= '3.6'
